+++ packages/demo-next/pages/blocks.tsx
@@ -62,6 +62,8 @@ export default function BlocksExample({ jsonFile }) {
               itemProps={{
                 style: { backgroundColor: 'red' },
               }}
+              min={2}
+              max={4}
             />
           </Wrap>
         </InlineForm>