+++ packages/@tinacms/react-core/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.3.2](https://github.com/tinacms/tinacms/compare/@tinacms/react-core@0.3.1...@tinacms/react-core@0.3.2) (2020-06-29)
+
+**Note:** Version bump only for package @tinacms/react-core
+
+
+
+
+
 ## [0.3.1](https://github.com/tinacms/tinacms/compare/@tinacms/react-core@0.3.1-alpha.0...@tinacms/react-core@0.3.1) (2020-06-23)